rkt, stage0, stage1: Add enter command
It should be used as follows:
rkt enter IMAGE [optional binary to run [optional parameters to binary]]

If no optional binary is provided then root shell inside container
will be executed.

Under the hood it is basically a hack - we call systemd-nspawn without
boot parameter and we don't create any unit files. Because of that we
need to pass different directory parameter to systemd-nspawn - namely
stage2 directory.