pkg: add an acirenderer library.
This library renders an aci given a well defined list of dependencies.
A future patch will also provide the dependency resolving logic.

It adds two interfaces: ACIProvider and ACIRegistry.

The first interface is the one used by this patch and the cas has been changed
to satisfy it.
The second interface will be used by the dependency resolution functions. Also
this interface is satisfied by the cas.