lock: add lock functions to lock by specific keys.
This adds a KeyLock struct and related functions trying to keep the same
semantic of DirLock.
It'll be useful for diskv file locking, for download locking and for every
need to lock something that isn't referenceable with a specific and already
available path on the filesystem.
By now it uses DirLock for the key lock files. From an aesthetics point of
view, a file (instead of an empty dir) will be better. This will need little
changes to DirLock.

KeyLock represents a lock for a specific key. The lock files (actually they
are directories as DirLock is used) are created inside a directory.
As a lock file is created for every key, there's the need to remove old lock
files (for example by a garbage collection function).
It's not possible to just remove a lock file taking an exclusive lock
on it as there can be other users waiting for a lock on it and if the file
is removed from the filesystem these users will have its fd open and will
continue to take/wait for a lock while new users will create a new lock file
creating a race condition.
For this reason the image lock file removal needs to be done taking an
exclusive lock on the directory containing the lock.