lock: add lock functions to lock by specific keys.
This adds a KeyLock struct and related functions trying to keep the same
semantic of DirLock.
It'll be useful for diskv file locking, for download locking and for every
need to lock something that isn't referenceable with a specific and already
available path on the filesystem.
By now it uses DirLock for the key lock files. From an aesthetics point of
view, a file (instead of an empty dir) will be better. This will need little
changes to DirLock.

KeyLock is a lock for a specific key. The locks files are created inside a
directory using the key name.
The key name must be a valid file name.