rkt: add support for --mounts to `rkt run` and `rkt prepare`
This introduces the concept of requiring app names with images on
the cli:

$  rkt run --mount app=test,volume=foo,target=work \
  --volume foo,source=/tmp,kind=host test=imgs/mounter.aci

There needs to be a way to address the app instances, here we just require
unique names be provided.

To keep things simple and consistent, the name is always required, even for
trivial runs:
$ rkt run simple=imgs/pauser.aci

Things are mostly unchanged, just a name= is expected before the image now,
and that name is referred to by --mount to specify which app the mount
pertains to.

This will probably be utilized by --set-env for setting app-specific
variables, and anything else we need to permit setting app-specifically.