*: use RuntimeApp.Name instead of Image.ID as app identifier
With the app names introduced for --mount, it also became possible to
include the same image multiple times in a single container, meaning
Image.ID is no longer guaranteed to be unique.

Instead use the app names for naming things like the stage2 per-app image
directories and systemd units.

This also changes the `rkt status` output to include the user-supplied app
names for the per-app status codes instead of the image ids, prefixed with
app=:
 $ /rkt status 04
 state=exited
 pid=23577
 exited=true
 app=sleepy=0
 $

which was run via:
 $ rkt run sleepy=imgs/sleeper.aci

This change depends on https://github.com/appc/spec/pull/247 landing in
appc/spec as well, which introduces helpers for (un)escaping types.ACName.