stage0: add ability to inject arbitrary volumes, refs coreos/rkt#761
Injected volumes is a runtime flag added to rkt tool with the ability to
mount arbitrary volumes in the container.

To inject an arbitrary volume when starting pod, do this:

rkt run app.aci --inject-volume=kind=host,source=/source/path,dest=/dest/path

Implementation details:

* Implement it as a part of stage0 without changing spec or other stages.
* Make sure the injected volume does not conflict with the volume defined in the Pod
to avoid errors.
* Generate and inject mount points into the pod spec during the prepare phase.
* Move the logic of overriding pod app during stage1 in case if image app is missing
to the prepare stage0 instead of stage1 phase to be able to specify the mount points