WIP: metadata-server: listen on a different unix socket
Instead of hard coding /run/rkt/metadata-svc.sock, rkt gains a new
option --run-dir=/run/rkt in addition to the existing --dir,
--local-config etc.

This allows to start the metadata-server for the tests in parallel to
another rkt metadata-server, or to run tests in parallel.

This is based on https://github.com/coreos/rkt/pull/1101 but it is not
finished. It does not work yet.