store: Split GetImageManifest into two functions
GetImageManifestJSON returns the manifest as bytes to be parsed. So if
this function returns an error, it means that the image was actually
removed.

GetImageManifest might return an error in two cases - either the image
was removed or its manifest was invalid.

The GetImageManifestJSON will be used in rkt image list.