stage1/reaper: make reaper to reap when each exits.
Previously, when an app exits with code == 0, its status will not
be written to file unless all apps exit.
On the other hand, if the exited app has non-zero exit code, the
whole pod will exit, and the exit status will be recorded.

This PR changes the above behavior that when a app exits, its exit
code will be recorded as soon as the app exits no matter it's zero
or not. Besides, the exit of the app will not cause the whole pod
to exit.