directory parameters: allow relative paths
The parameters --dir=, --system-config= and --local-config= used to only
accept absolute paths. This patch changes the type of the parameters
from string to absDir. When the flag is parsed, the potentially relative
path is converted to an absolute path automatically.

The parameters still have the same defaults. It is no longer allowed to
pass an empty string (e.g. --dir="" is not allowed).

The value can be accessed using "globalFlags.Dir.String()". This makes
this patch big because I had to add the ".String()" suffix each time the
values are accessed.

Fixes https://github.com/coreos/rkt/issues/1605