stage0: added support for fetching plugins, refactored docker/http(s)
Adds support for having images fetched via a plugin. Plugins take the
form of a binary on the user's path named "rkt-fetcher-%s", with the %s
taking the place of the schema, so for example fetching
`docker://alpine` is redirected to the `rkt-fetcher-docker` binary.

When the plugin is started it has a JSON configuration written to it on
stdin, its stderr is forwarded on to the user, and its expected to write
result json to stdout. The configuration contains paths to write the ACI
and ASC files to, which rkt will then verify (unless the insecure image
option is specified).