xen: introduce run, enter, stop
Introduce simple bash scripts to implement "run", "enter" and "stop" for
the xen flavor. They rely on the "xl" command line tool for doing the
work.

The run script generates an xl config file on the fly based on the
command line options and stage2 manifest.

The enter script finds the domid of the VM, then calls "xl console".

The stop script destroys the VM.

Signed-off-by: Stefano Stabellini <stefano@aporeto.com>