stage1/units: introduce a new journal logging mode
It is similar to the existing `log` mode that forwards all logs to
systemd, with the difference that output will not show up on the console
as well (StanddardOutput|Error=journal instead of journal+console).

In addition to that, a read-only view of the pod journal will be bind
mounted inside each app, so one more apps can decide to forward output
from the pod reading that journal.

Signed-off-by: Fabio Kung <fabio.kung@gmail.com>