write /etc/machine-id for each app
... in their respective stage2/rootfs/etc/machine-id paths.

This allows apps to read the Pod journal when it is available.
systemd has a similar mechanism with systemd-firstboot.service and
systemd-machine-id-setup(1). The stage1 image could use that service,
since nspawn is already passing the container_uuid[1] ENV var, but
writing the machine-id files directly was simpler than bringing in a new
systemd service/dependency.

[1]: https://www.freedesktop.org/wiki/Software/systemd/ContainerInterface/

Signed-off-by: Fabio Kung <fabio.kung@gmail.com>