stage1/units/lifecycle: allow graceful shutdown on immutable pods
Conflicts+Before: halt.target causes app units to be stopped in the
correct phase by systemd, allowing KillMode and TimeoutStopSec to be
in effect.

This also adds After dependencies for the reaper and shutdown, to ensure
the reaper runs only after the app has been gracefully terminated, and
to prevent abruptly terminating journald when the halt.target gets
activated (e.g. by rkt stop).

The default KillMode is `control-group`, and the default timeout before
sending SIGKILLs is 90s. Which means systemd will broadcast SIGTERMs to
all processes in the app, wait at most 90s for all processes to
terminate, then finish the remaining with SIGKILLs.

The default timeout is documented on systemd-system.conf(5). The default
KillMode is on systemd.kill(5).

A followup commit will allow the above to be customizable per app.

Signed-off-by: Fabio Kung <fabio.kung@gmail.com>