experimental: apparmor confinement support
This is a hidden experimental feature, that allows apps to specify an
AppArmor profile to be confined with in a Pod manifest, via a rkt
specific annotation.

Only custom built stage1 images (e.g.: src flavor compiled with apparmor
support, or host flavor with a systemd on the host that has apparmor
enabled) will support this annotation, and it will be a noop on all
stage1 images bundled with rkt.

This also adds a new `--insecure-options=apparmor` to completely ignore
AppArmor annotations in Pod manifests, which turn them onto a noop even
on stage1 images that support apparmor.

Signed-off-by: Fabio Kung <fkung@netflix.com>